feat: add wrappers and quantum utilities tests
