+++ dashboard/README.md
@@ -25,7 +25,8 @@ This module is designed to meet enterprise auditability and compliance standards
 
 ## ARCHITECTURE
 
-- **Backend:** Flask application (`dashboard/enterprise_dashboard.py`)
+- **Backend:** Flask application (`dashboard/enterprise_dashboard.py` wraps
+  `web_gui.scripts.flask_apps.enterprise_dashboard`)
 - **Templates:** Jinja2 HTML (`dashboard/templates/`)
 - **Static Content:** CSS, JS, images (`dashboard/static/`)
 - **Data Sources:** `production.db`, `analytics.db`, `monitoring.db`
@@ -71,7 +72,7 @@ dashboard/
 ### Starting the Dashboard
 
 ```bash
-python dashboard/enterprise_dashboard.py
+python dashboard/enterprise_dashboard.py  # wrapper for web_gui Flask app
 ```
 
 Visit [http://localhost:5000](http://localhost:5000) in your browser. The dashboard will auto-discover and display current session, database, and compliance data. All metrics update in real time.
@@ -94,6 +95,22 @@ Visit [http://localhost:5000](http://localhost:5000) in your browser. The dashbo
 #### Example `/dashboard/compliance` Response
 
 The `/dashboard/compliance` endpoint returns compliance information as JSON, combining live metrics from `analytics.db` and correction/rollback summaries from `dashboard/compliance/correction_summary.json`.
+
+`metrics.json` uses the following schema:
+
+```json
+{
+  "metrics": {
+    "placeholder_removal": 0,
+    "compliance_score": 0.0,
+    "violation_count": 0,
+    "rollback_count": 0,
+    "last_update": "ISO8601 timestamp"
+  },
+  "status": "updated",
+  "timestamp": "ISO8601 timestamp"
+}
+```
 
 ```json
 {
