+++ dashboard/compliance/correction_summary.json
@@ -1,1 +1,14 @@
-{"corrections": [], "timestamp": "placeholder"}
+{
+  "timestamp": "2025-07-25T03:44:32.983024",
+  "total_corrections": 1,
+  "corrections": [
+    {
+      "file_path": "/tmp/workspace/example.txt",
+      "rationale": "Corrected syntax error in example.txt",
+      "compliance_score": 0.95,
+      "rollback_reference": "/tmp/workspace/backups/example.txt.bak",
+      "timestamp": "2025-07-25T03:44:32.952002"
+    }
+  ],
+  "status": "complete"
+}
