+++ dashboard/compliance/metrics.json
@@ -1,1 +1,11 @@
-{"metrics": {}, "status": "placeholder"}
+{
+  "metrics": {
+    "placeholder_removal": 0,
+    "compliance_score": 0.0,
+    "violation_count": 0,
+    "rollback_count": 0,
+    "last_update": "2025-07-25T03:43:39.942598"
+  },
+  "status": "updated",
+  "timestamp": "2025-07-25T03:43:39.946461"
+}
