+++ scripts/unified_legacy_cleanup_system.py
@@ -16,6 +16,7 @@ from typing import List
 
 from db_tools.core.connection import DatabaseConnection
 from db_tools.core.models import DatabaseConfig
+from utils.log_utils import _log_event
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO, format="%(message)s")
@@ -62,7 +63,9 @@ class UnifiedLegacyCleanupSystem:
         try:
             rows = self.file_conn.execute_query(query)
             scripts = [self.config.workspace_path / r["script_path"] for r in rows]
-            return [s for s in scripts if s.exists()]
+            valid_scripts = [s for s in scripts if s.exists()]
+            _log_event({"event": "discover_legacy_scripts", "count": len(valid_scripts)})
+            return valid_scripts
         except Exception as exc:  # pragma: no cover - database might not exist
             logger.warning(f"Database query failed: {exc}")
             return []
@@ -80,9 +83,11 @@ class UnifiedLegacyCleanupSystem:
             return True
         try:
             script.rename(target)
+            _log_event({"event": "archive_script", "script": str(script)})
             return True
         except Exception as exc:  # pragma: no cover - file system errors
             logger.error(f"Archive failed: {exc}")
+            _log_event({"event": "archive_failed", "script": str(script), "error": str(exc)})
             return False
 
     def optimize_workspace(self, dry_run: bool = False) -> None:
@@ -90,7 +95,9 @@ class UnifiedLegacyCleanupSystem:
         logger.info("Optimizing workspace layout")
         if dry_run:
             return
+        _log_event({"event": "optimize_workspace_start"})
         # Currently a no-op; real implementation would reorganize files
+        _log_event({"event": "optimize_workspace_complete"})
         return
 
     def run_cleanup(self, dry_run: bool = False) -> None:
